requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
slowapi
chromadb
sentence-transformers
numpy
google-generativeai
pypdf
python-dotenv
